Fix attack priority and health bar bugs
Fixed an annoying bug where P1 would always get priority if both
players attack on the same frame. This was basically caused by the
updating of hitboxes being handled by Unity's animation manager, which
was probably running on a different thread than the one which updates
the players. Now Unity's animation merely queues the updating of the
hitboxes, it ultimately happens predictably in the Player's Update() method

Also changed the behaviour of the health bar in the UI. Now the red
will appear to accumulate when hitting an opponent while they are
already in hitstun (comboing them). Also, if the red bar is depleting,
and the player is hit, the red will instantly snap to the position of
the green bar before taking the damage.