Add ko mechanics and round end logic
Now you can KO your opponent by depleting their life to zero.

When someone gets KO'd, this happens:
1. Time stops & "KO" or "Double KO" is displayed
2. Time goes at slow speed while player(s) is falling to ground
3. Time speeds back up to normal speed
4. Winner (if there is one) does victory pose

Still need to implement time-outs, round counting, restarting and overall
victory by winnig rounds.

The static `Globals` class is removed and is replaced with an instance
of a new `Globals` class which inherits from `ScriptableObject`. This
instance is in `Assets/Resources` and is generally accessed the way
resources are normally accessed with `Resources.Load()`